requirements.txt
nibabel>=5.0
numpy>=1.24
pandas>=2.0
scipy>=1.11
matplotlib>=3.7
pyyaml>=6.0
nnunetv2
pytest>=7.0